Rename TabList, add styles, replace getNavItems with a new logic
